Provide O(1) access to netcdf-4 objects via name or external id
This PR implements a pervasive change to remove (almost) all
occurrences of linear search for various kinds of objects:
dimensions, types, groups, variables, and attributes.
Linear search is replaced with some combination of hashing
(for name-based lookup) and/or vectors (for numeric id-based
lookup).

The change is described in detail in the file docs/indexing.dox,
which is can be viewed if the "--enable-internal-docs" flag is passed
to ./configure.